move pipeline upload script to express common
