--- examples/pipelines/finetune_stable_diffusion/helpers/lightweight.py
@@ -1,12 +1,0 @@
-"""Collection of lightweight components
-"""
-
-
-# example lightweight component definition
-def read_input(input_v: str):
-    """Lightweight component to print outputs from other components
-    Args:
-        input_v (str): string that is to be printed out
-    """
-
-    print(input_v)
